Fix definition of output variables in get-existing-environment-info action

+++ .scripts/get-existing-environment-info.sh
@@ -3,18 +3,15 @@ source "$(dirname -- "${BASH_SOURCE[0]}")/functions/get-existing-environment-inf
 
 shopt -s expand_aliases
 
+output_arr="$1"
+
 test -n "$ACCOUNT"      || { echo "Variable 'account' missing"; exit 1; }
 test -n "$ENVIRONMENT"  || { echo "Variable 'environment' missing"; exit 2; }
 test -f "$CONFIG_PATH"  || { echo "Config '$CONFIG_PATH' file not found"; exit 3; }
+test -n "$output_arr"   || { echo "Variable 'output_arr' missing"; exit 4; }
 
 alias yq='docker run --rm -v $PWD:/workdir mikefarah/yq'
 
-declare -A info
+declare -n output="$output_arr"
 
-get-existing-environment-info info "$ACCOUNT" "$ENVIRONMENT" "$CONFIG_PATH"
-
-for key in "${!info[@]}"
-do
-  echo "::set-output name=$key::${info[$key]}"
-done
-
+get-existing-environment-info output "$ACCOUNT" "$ENVIRONMENT" "$CONFIG_PATH"
